forum overall added field active to deprecate status field

+++ plugins/forum/share/db_installer/sql/forum_reports.sql.php
@@ -5,6 +5,6 @@ $data = '
 	`user_id` int(11) NOT NULL default \'0\',
 	`time` int(11) NOT NULL default \'0\',
 	`text` text NOT NULL,
-	`active` int(11) NOT NULL default \'1\',
+	`active` tinyint(1) NOT NULL default \'1\',
 	PRIMARY KEY  (`id`)
 ';
